% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aridge_solver}
\alias{aridge_solver}
\title{Fit B-splines with automatic knot selection.}
\usage{
aridge_solver(
  x,
  y,
  knots = seq(min(x), max(x), length = 42)[-c(1, 42)],
  pen = 10^seq(-3, 3, length = 100),
  degree = 3L,
  maxiter = 1000,
  epsilon = 1e-05,
  verbose = FALSE,
  tol = 1e-06
)
}
\arguments{
\item{x}{Data x values}

\item{y}{Data y values}

\item{knots}{Knots}

\item{pen}{A vector of positive penalty values. The adaptive spline regression is performed for every value of pen}

\item{degree}{The degree of the splines. Recommended value is 3, which corresponds to natural splines.}

\item{maxiter}{Maximum number of iterations  in the main loop.}

\item{epsilon}{Value of the constant in the adaptive ridge procedure (see \emph{Frommlet, F., Nuel, G. (2016)
An Adaptive Ridge Procedure for L0 Regularization}.)}

\item{verbose}{Whether to print details at each step of the iterative procedure.}

\item{tol}{The tolerance chosen to diagnostic convergence of the adaptive ridge procedure.}

\item{diff}{Order of the differences on the parameters. The value \code{degree + 1} is necessary to perform
selection of the knots.}
}
\description{
Fit B-splines with automatic knot selection.
}
