% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aspline}
\alias{aspline}
\title{Fit A-splines}
\usage{
aspline(x, y, degree, n_knots = 50)
}
\arguments{
\item{x}{Numerical vector, data x values. Must be sorted in increasing order.}

\item{y}{Numerical vector, data y values.}

\item{degree}{An integer greater or equal to zero. The degree of the spline (degree = order + 1).}

\item{n_knots}{Maximum number of knots to consider.}
}
\description{
Fit A-splines
}
