% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{band_weight}
\alias{band_weight}
\title{Create the matrix of weights}
\usage{
band_weight(w, diff)
}
\arguments{
\item{w}{The vector of weights}

\item{diff}{The order of the differences to be applied to the parameters. Must be a strictly positive integer}
}
\value{
The matrix of weights \eqn{D^T diag(w) D} where
 \code{D <- diff(diag(length(w) + diff), differences = diff)}. Only the non-null superdiagonals of
 the weight matrixare returned, each column corresponding to a diagonal.
}
\description{
Create the matrix of weights
}
