% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bladder}
\alias{bladder}
\title{Bladder Cancer aCGH profile data}
\format{
A data frame with 500 observations and 2 variables:
\describe{
  \item{x}{probe number}
  \item{y}{aCGH profile value}
}
}
\source{
Stransky, N., Vallot, C., Reyal, F., Bernard-Pierrot, I., de Medina, S. G. D., Segraves,
R., de Rycke, Y., Elvin, P., Cassidy, A., Spraggon, C., Graham, A., Southgate, J.,
Asselain, B., Allory, Y., Abbou, C. C., Albertson, D. G., Thiery, J. P., Chopin, D. K.,
Pinkel, D. and Radvanyi, F. (2006). Regional Copy Number Independent Deregulation
of Transcription in Cancer', \emph{Nature Genetics} 38(12), 1386-1396.
}
\usage{
bladder
}
\description{
A dataset of 500 observations corresponding to 500 probes of the aCGH profile of a bladder cancer patient.
The original data are provided by \cite{Stransky et al. (2006)}. This dataset consists of probes 1 through 500 of individual 1.
}
\keyword{datasets}
