% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{block_design}
\alias{block_design}
\title{Transform a Spline Design Matrix in block compressed form}
\usage{
block_design(X, degree)
}
\arguments{
\item{X}{The design matrix, as given by \code{splines2::bSpline}.}

\item{degree}{Degree of the spline regression, as used in function \code{splines2::bSpline}.}
}
\value{
A matrix \code{B} with all non-zero entries of \code{X} and a vector of indices \code{alpha} representing the positions of the non-zero blocks of \code{X}.
}
\description{
Transform a Spline Design Matrix in block compressed form
}
