% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fossil}
\alias{fossil}
\title{Fossil data}
\format{
A data frame with 106 observations and 2 variables:
\describe{
  \item{age}{The age of fossils, in millions of years}
  \item{strontium.ratio}{Ratio of strontium isotopes}
  ...
}
}
\source{
Bralower, T.T, Fullagar, P.D., Paull, C.K, Dwyer, G.S. and Leckie, R.M. (1997). Mid-cretaceous strontium-isotope stratigraphy of deep-sea sections. \emph{Geological Society of America Bulletin}, 109, 1421-1442.
}
\usage{
fossil
}
\description{
A dataset with 106 observations on fossil shells from the \code{SemiPar} package (\url{https://CRAN.R-project.org/package=SemiPar}).
}
\references{
Ruppert, D., Wand, M.P. and Carroll, R.J. (2003). \emph{Semiparametric Regression}, Cambridge University Press.
}
\keyword{datasets}
