% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LDL}
\alias{LDL}
\title{LDL}
\arguments{
\item{D}{Rotated row-wised matrix of dimensions n*k, with first column corresponding to the diagonal, the second to the first super-diagonal and so on.}
}
\value{
List with D as solution of our LDL decomposition.
}
\description{
Fast inplace LDL decomposition of symmetric band matrix of length k.
}
\examples{

n=10;
D0=1:10;
D1=exp(-c(1:9));
D=cbind(D0,c(D1,0))
sol=LDL(D)
}
