% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lidar}
\alias{lidar}
\title{Lidar data}
\format{
\describe{
   \item{range}{distance travelled before the light is reflected back to its source}
   \item{logratio}{logarithm of the ratio of received light from two laser sources}}
}
\source{
\itemize{
\item Sigrist, M. (Ed.) (1994). Air Monitoring by Spectroscopic Techniques (Chemical Analysis Series,vol. 197). New York: Wiley
\item The R package \url{https://CRAN.R-project.org/package=SemiPar}}
}
\usage{
lidar
}
\description{
Data from a light detection and ranging (LIDAR) experiment
}
\references{
Ruppert, D., Wand, M.P. and Carroll, R.J. (2003). \emph{Semiparametric Regression}, Cambridge University Press.
}
\keyword{datasets}
