% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{montreal}
\alias{montreal}
\title{Montreal Temperature Data}
\format{
A data frame with 730 rows and 2 variables:
\describe{
  \item{day}{The day of the year from January 1, 1961, to December 31, 1962}
  \item{temp}{Temperature in Celsius}
  ...
}
}
\source{
\code{fda::"MontrealTemp"}
}
\usage{
montreal
}
\description{
A dataset containing the tempature in Montreal for two years
}
\keyword{datasets}
