% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nmr}
\alias{nmr}
\title{Nuclear Magnetic Resonance data}
\format{
Data farme of 1024 rows and two columns: the index \code{x} and the signal \code{y}.
}
\source{
\itemize{
\item Data from \url{https://web.stanford.edu/~hastie/ElemStatLearn/datasets/nmr1.csv}.
\item See also The Elements of Statisical Learning (2001, 2nd Ed.), \emph{Hastie, T., Friedman, J., and Tibshirani, R.J}, p. 176}.
}
\usage{
nmr
}
\description{
A signal of nuclear magnetic resonance.
}
\keyword{datasets}
