% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanium}
\alias{titanium}
\title{Titanium heat data}
\format{
49 observations and two variables:
\describe{
\item{x}{temperature}
\item{y}{physical property}
}
}
\source{
\itemize{
\item de Boor, C., and Rice, J. R. (1986), Least-squares cubic spline approximation. II: variable knots. \emph{Report CSD TR 21, Purdue U., Lafayette, IN.}
\item Dierckx, P. (1993), \emph{Curve and Surface Fitting with Splines}, Springer.
\item Jupp, D. L. B. (1975), \emph{Approximation to data by splines with free knots}, SIAM Journal on Numerical Analysis, 15: 328-343.
}
}
\usage{
titanium
}
\description{
A data set of 49 samples expressing the thermal property of titanium
}
\keyword{datasets}
