% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{wridge_solver_fast_2}
\alias{wridge_solver_fast_2}
\title{Fit B-Splines with weighted penalization over differences of parameters}
\usage{
wridge_solver_fast_2(XX_band, Xy, degree, pen, w = rep(1, nrow(XX_band) -
  degree - 1), old_par = rep(1, nrow(XX_band)), maxiter = 1000,
  tol = 1e-08)
}
\arguments{
\item{degree}{The degree of the B-splines.}

\item{pen}{Positive penalty constant.}

\item{w}{Vector of weights. The case \eqn{\mathbf w = \mathbf 1} corresponds to fitting P-splines with difference #' order \code{degree + 1} (see \emph{Eilers, P., Marx, B. (1996) Flexible smoothing with B-splines and penalties}.)}

\item{old_par}{Initial parameter to serve as starting point of the iterating process.}

\item{maxiter}{Maximum number of Newton-Raphson iterations to be computed.}
}
\value{
The estimated parameter of the spline regression.
}
\description{
Fit B-Splines with weighted penalization over differences of parameters
}
