% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{aspline}
\alias{aspline}
\alias{aridge_solver}
\title{Fit B-splines with automatic knot selection.}
\usage{
aspline(
  x,
  y,
  knots = seq(min(x), max(x), length = 42)[-c(1, 42)],
  pen = 10^seq(-3, 3, length = 100),
  degree = 3L,
  family = c("gaussian"),
  maxiter = 1000,
  epsilon = 1e-05,
  verbose = FALSE,
  tol = 1e-06
)

aridge_solver(
  x,
  y,
  knots = seq(min(x), max(x), length = 42)[-c(1, 42)],
  pen = 10^seq(-3, 3, length = 100),
  degree = 3L,
  family = c("gaussian"),
  maxiter = 1000,
  epsilon = 1e-05,
  verbose = FALSE,
  tol = 1e-06
)
}
\arguments{
\item{x, y}{Input data, numeric vectors of same length}

\item{knots}{Knots}

\item{pen}{A vector of positive penalty values. The adaptive spline regression is performed for every value of pen}

\item{degree}{The degree of the splines. Recommended value is 3, which corresponds to natural splines.}

\item{family}{For forward compatibility. A description of the error distribution and link function to be used in the model. Only the "gaussian" family is implemented yet, corresponding to linear regression.}

\item{maxiter}{Maximum number of iterations  in the main loop.}

\item{epsilon}{Value of the constant in the adaptive ridge procedure (see \emph{Frommlet, F., Nuel, G. (2016)
An Adaptive Ridge Procedure for L0 Regularization}.)}

\item{verbose}{Whether to print details at each step of the iterative procedure.}

\item{tol}{The tolerance chosen to diagnostic convergence of the adaptive ridge procedure.}
}
\description{
Fit B-splines with automatic knot selection.
}
\section{Functions}{
\itemize{
\item \code{aridge_solver}: Alias for aspline, for backwards compatibility
}}

